pyproject.toml
[tool.poetry]
name = "leadanne2"
version = "0.2.0"
description = ""
authors = ["Pedro Antonio <pedro@z33dd.com>"]
readme = "README.md"

[tool.poetry.scripts]
leadanne2 = "leadanne2.cli:app"

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.34.0"
python-dotenv = "^1.0.0"
kombu = "^5.3.4"
sentry-sdk = "^1.39.1"
langchain = "^0.1.5"
langchain-openai = "^0.0.5"
celery = "^5.3.6"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
typer = "^0.9.0"
langchain-community = "^0.0.27"
playwright = "^1.43.0"
bs4 = "^0.0.2"
supabase = "^2.4.5"
jinja2 = "^3.1.4"
resend = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
ipykernel = "^6.29.0"
pypdf = "^4.0.1"
pytest = "^8.2.0"
pytest-snapshot = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
